MIN: Moving create_local_model_class to test_utils/models and making add_mixin public.
