+++ src/porepy/applications/test_utils/models.py
@@ -306,7 +306,7 @@ def subdomains_or_interfaces_from_method_name(
     return domains
 
 
-def _add_mixin(
+def add_mixin(
     mixin: type[pp.PorePyModel], parent: type[pp.PorePyModel]
 ) -> type[pp.PorePyModel]:
     """Helper method to dynamically construct a class by adding a mixin.
@@ -326,6 +326,23 @@ def _add_mixin(
     # such an addition could not be made in the mixin class instead.
     cls = type(name, (mixin, parent), {})
     return cls
+
+
+def create_local_model_class(
+    model_class: type[pp.PorePyModel], mixin_models: list[type[pp.PorePyModel]]
+) -> type[pp.PorePyModel]:
+    """Helper method to add mixins to a model class for testing purpose.
+
+    Note that the order in ``mixin_models`` has an effect. First element will be mixed
+    in first, last element last.
+
+    """
+    local_model_class: type[pp.PorePyModel] = model_class
+
+    for mixin_model in mixin_models:
+        local_model_class = add_mixin(mixin_model, local_model_class)
+
+    return local_model_class
 
 
 def compare_scaled_primary_variables(
