+++ tests/models/test_geometry.py
@@ -30,11 +30,11 @@ from porepy.applications.test_utils.arrays import projection_matrix_from_array_s
 # List of geometry classes to test.
 # Turn mixins of specific grids into proper model geometries.
 geometry_list: list[type[pp.ModelGeometry]] = [
-    models._add_mixin(
+    models.add_mixin(
         porepy.applications.md_grids.model_geometries.RectangularDomainThreeFractures,
         pp.ModelGeometry,
     ),
-    models._add_mixin(
+    models.add_mixin(
         porepy.applications.md_grids.model_geometries.OrthogonalFractures3d,
         pp.ModelGeometry,
     ),
